Rename claimManager dir to primitives
