+++ client/primitives/getNewQueryStep.js
