+++ client/primitives/triggerVerificationGame.js
